refactor: extract IED search filter into its own component file
Stored in src/components/ied-search-filter.ts
